Backend/sunrise/requirement.txt
django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
PyJWT==1.7.1 
djangorestframework-jwt==1.11.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
